+++ lib/src/delegates/asset_picker_builder_delegate.dart
@@ -242,11 +242,7 @@ abstract class AssetPickerBuilderDelegate<Asset, Path> {
 
   /// Determine how to browse assets in the viewer.
   /// 定义如何在查看器中浏览资源
-  Future<void> viewAsset(
-    BuildContext context,
-    int index,
-    AssetEntity currentAsset,
-  );
+  Future<void> viewAsset(BuildContext context, int? index, Asset currentAsset);
 
   /// Yes, the build method.
   /// 没错，是它是它就是它，我们亲爱的 build 方法~
@@ -870,7 +866,7 @@ class DefaultAssetPickerBuilderDelegate
   @override
   Future<void> viewAsset(
     BuildContext context,
-    int index,
+    int? index,
     AssetEntity currentAsset,
   ) async {
     final DefaultAssetPickerProvider provider =
@@ -902,6 +898,10 @@ class DefaultAssetPickerBuilderDelegate
         selected = provider.selectedAssets;
         effectiveIndex = current.indexOf(currentAsset);
       }
+    } else if (index == null) {
+      current = provider.selectedAssets;
+      selected = provider.selectedAssets;
+      effectiveIndex = selected.indexOf(currentAsset);
     } else {
       current = provider.currentAssets;
       selected = provider.selectedAssets;
@@ -1936,7 +1936,7 @@ class DefaultAssetPickerBuilderDelegate
       child: Consumer<DefaultAssetPickerProvider>(
         builder: (context, DefaultAssetPickerProvider p, __) => GestureDetector(
           onTap: p.isSelectedNotEmpty
-              ? () => viewAsset(context, 0, p.selectedAssets.first)
+              ? () => viewAsset(context, null, p.selectedAssets.first)
               : null,
           child: Selector<DefaultAssetPickerProvider, String>(
             selector: (_, DefaultAssetPickerProvider p) =>