+++ CHANGELOG.md
@@ -13,6 +13,11 @@ that can be found in the LICENSE file. -->
 
 - Use `wechat_picker_library`.
 
+### Fixes
+
+- Fix previewing selected assets behavior.
+- Use `PermissionRequestOption` as much as possible.
+
 ## 9.0.0-dev.2
 
 ### Fixes